Separate WaitGroup (#572)
* Separate WaitGroup

* Update core/utils/wait_group.hpp


* Exact count

* fix lints

---------
