+++ core/index/index_writer.hpp
@@ -45,6 +45,7 @@
 #include "utils/object_pool.hpp"
 #include "utils/string.hpp"
 #include "utils/thread_utils.hpp"
+#include "utils/wait_group.hpp"
 
 #include <absl/container/flat_hash_map.h>
 
@@ -835,36 +836,7 @@ class IndexWriter : private util::noncopyable {
     std::deque<PendingSegmentContext> pending_segments_;
     // entries from 'pending_segments_' that are available for reuse
     Freelist pending_freelist_;
-    // TODO(MBkkt) Considered to replace with YACLib in ArangoDB 3.11+ or ...
-    struct WaitGroup {
-      std::mutex& Mutex() noexcept { return m_; }
-
-      void Add() noexcept { counter_.fetch_add(1, std::memory_order_relaxed); }
-
-      void Done() noexcept {
-        if (counter_.fetch_sub(1, std::memory_order_acq_rel) == 1) {
-          std::lock_guard lock{m_};
-          cv_.notify_one();
-        }
-      }
-
-      void Wait(std::unique_lock<std::mutex>& lock) noexcept {
-        IRS_ASSERT(lock.mutex() == &m_);
-        IRS_ASSERT(lock.owns_lock());
-        if (counter_.fetch_sub(1, std::memory_order_acq_rel) != 1) {
-          do {
-            cv_.wait(lock);
-            // relaxed probably enough
-          } while (counter_.load(std::memory_order_acquire) != 0);
-        }
-        counter_.store(1, std::memory_order_relaxed);
-      }
-
-     private:
-      std::atomic_size_t counter_{1};
-      std::mutex m_;
-      std::condition_variable cv_;
-    } pending_;
+    WaitGroup pending_;
 
     // set of segments to be removed from the index upon commit
     ConsolidatingSegments segment_mask_;
